Move hook domain logic from routes to domain (#1774)
Moves domain-considered logic from `CacheHooksService` to a new `HooksRepository` in the domain:

- Move domain logic from `CacheHooksService` to `HooksRepository`
- Propagate changes accordingly
